Split search modules and repair validation gates
